backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
psycopg==3.3.2
python-dotenv==1.2.1
pydantic==2.12.5
asyncpg==0.29.0
